feature/MSC-01: Party movement is now configurable, actor movement is stored in floats in database now, removed previous animation config as no longer neccesary

+++ Kurumi-Engine/Data/Runtime/Actors/Core/Actor.cs
@@ -77,7 +77,7 @@ public class Actor : IFacingPositionProvider, IMapEntity, IMutablePositionProvid
     // Actor movement functions.
     public int SpriteId => _actorInfo.SpriteId;
 
-    public int MovementSpeed => _actorInfo.MovementSpeed;
+    public float MovementSpeed => _actorInfo.MovementSpeed;
 
     internal Actor(ActorInfo actorInfo, ActorModel actorModel)
     {
